requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2024.12.14
click==8.1.8
distro==1.9.0
Flask==3.1.0
Flask-CORS==4.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
openai==1.60.2
pymongo==4.6.2
PyJWT==2.8.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
bcrypt==4.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-speech==2.23.0
google-cloud-storage==2.12.0
google-cloud-aiplatform==1.38.0
pydub==0.25.1
ffmpeg-python==0.2.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
google-cloud-firestore==2.13.1
google-cloud-secretmanager==2.18.1
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0
google-auth==2.25.2
sendgrid==6.10.0
pydantic[email]==2.5.0
email-validator==2.1.0

